pyproject.toml
[tool.poetry]
name = "nomad-django"
version = "0.1.0"
description = ""
authors = ["ggomdong <ggomdong@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
